Get rid of special instructions for spawn and halt
