+++ context.ts
@@ -1,5 +1,5 @@
 import { Context, Delimiter, Operation } from "./types.ts";
-import { Do } from "./control.ts";
+import { Do, Resume } from "./control.ts";
 import { Ok } from "./result.ts";
 import { getContext, setContext } from "./-context.ts";
 
@@ -7,11 +7,11 @@ export function createContext<T>(name: string, defaultValue?: T): Context<T> {
   let context: Context<T> = { name, get, set, expect, defaultValue };
 
   function* get(): Operation<T | undefined> {    
-    return (yield Do((coroutine) => Ok(getContext(context, coroutine)))) as T | undefined;
+    return (yield Do((routine) => routine.next(Resume(Ok(getContext(context, routine)))))) as T | undefined;
   }
 
   function* set(value: T): Operation<T> {
-    return (yield Do((routine) => Ok(setContext(context, routine, value)))) as T;
+    return (yield Do((routine) => routine.next(Resume(Ok(setContext(context, routine, value)))))) as T;
   }
 
   function* expect(): Operation<T> {
